Use long for HLS media sequences
Issue:#3747

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183366339

+++ library/hls/src/main/java/com/google/android/exoplayer2/source/hls/playlist/HlsPlaylistTracker.java
@@ -450,9 +450,11 @@ public final class HlsPlaylistTracker implements Loader.Callback<ParsingLoadable
 
   private static Segment getFirstOldOverlappingSegment(HlsMediaPlaylist oldPlaylist,
       HlsMediaPlaylist loadedPlaylist) {
-    int mediaSequenceOffset = loadedPlaylist.mediaSequence - oldPlaylist.mediaSequence;
+    long mediaSequenceOffset = loadedPlaylist.mediaSequence - oldPlaylist.mediaSequence;
     List<Segment> oldSegments = oldPlaylist.segments;
-    return mediaSequenceOffset < oldSegments.size() ? oldSegments.get(mediaSequenceOffset) : null;
+    return mediaSequenceOffset < oldSegments.size()
+        ? oldSegments.get((int) mediaSequenceOffset)
+        : null;
   }
 
   /**
